move util files to utils folder
